# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job API",
        "description": "API Documentation for Job Management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/cv/upload": {
            "post": {
                "tags": [
                    "CV"
                ],
                "summary": "Upload un CV",
                "operationId": "4257b56f20d4c62cd54e91e5ac87af78",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "cv": {
                                        "description": "Fichier CV à uploader",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CV uploadé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "CV uploaded successfully!"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "cvs/mon_cv.pdf"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur d'upload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Upload failed"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "File format not allowed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/cv/getcv": {
            "get": {
                "tags": [
                    "CV"
                ],
                "summary": "Récupérer les CV de l'utilisateur",
                "operationId": "29de2f89c708fdb89d12888c35bcff62",
                "responses": {
                    "200": {
                        "description": "Liste des CV",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "file_path": {
                                                "type": "string",
                                                "example": "cvs/mon_cv.pdf"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-03-18T12:34:56Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}